+++ indico/web/client/js/utils/date.js
@@ -7,6 +7,7 @@
 
 import _ from 'lodash';
 import moment from 'moment';
+import 'moment-timezone';
 
 export function toMoment(dt, format, strict = false) {
   if (!dt) {